# api.yaml
swagger: "2.0.0"
info:
  title: Academy APP
  description: Api Documentation for Academia Cyber Flow
  version: 1.0.0
  contact:
    email: choudharysidharth082000@gmail.com
  servers:
    - url: "http://localhost:3000"
      description: "Local Server Running"
    - url: "https://testcyberflow.herokuapp.com"
      description: "Main Server"
schemes: [http, https]
paths:
...
  /v1/userTopic/getAllTopicsByCourse/{userID}/{courseID}:
    get:
      tags:
        - name: User Course
      summary: Getting all User courses topics
      description: APi will get all the user courses in the database
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - name: courseID
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
...
  /v1/userTopic/startTopic/{userID}/{courseID}/{topicID}:
    post:
      tags:
        - name: User Course
      summary: Update only percentage
      description: APi will get all the user courses in the database
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - name: courseID
          in: path
          type: string
          required: true
        - name: topicID
          in: path
          type: string
          required: true
        - name: progress
          in: body
          schema:
            type: object
            properties:
              progress:
                type: number
      responses:
        200:
          description: OK
...
  /v1/timeline/{userID}/:
    get:
      tags:
        - name: Timeline
      summary: Timeline of a user
      parameters:
        - name: userID
          in: path
          type: string
          required: true
        - name: date
          in: query
          type: string
          required: true
          example: yyyy-mm-dd
      responses:
        200:
          description: OK
...
  /v1/userTimeline/add:
    post:
      tags:
        - timeline
      summary: add
      parameters:
        - in: body
          name: cart
          required: true
          schema:
            type: object
            properties:
              userID:
                type: string
              allday:
                type: boolean
              title:
                type: string
              description:
                type: string
              start:
                type: string
                format: date
                description: event start date
                example: "2021-01-30"
              end:
                type: string
                format: date
                description: end
                example: "2021-01-30"
              textColor:
                type: string
      responses:
        200:
          description: OK
  /v1/userTimeline/deleteUserTimeline/{timeLineId}:
    delete:
      tags:
        - timeline
      summary: delete
      parameters:
        - in: path
          name: timelineId
          type: string
          required: true
      responses:
        200:
          description: OK
  /v1/userTimeline/{userID}/:
    get:
      tags:
        - timeline
      summary: Timeline of a user
      parameters:
        - in: path
          name: userID
          type: string
          required: true
        - in: query
          name: date
          type: string
          required: true
          example: yyyy-mm-dd
      responses:
        200:
          description: OK
  /v1/userTimeline/all/{userID}/:
    get:
      tags:
        - timeline
      summary: Timeline of a user
      parameters:
        - in: path
          name: userID
          type: string
          required: true
      responses:
        200:
          description: OK